Move Partitioner to its own file.
