+++ Wrappers/Python/cil/framework/__init__.py
@@ -33,4 +33,4 @@ from .framework import AX, PixelByPixelDataProcessor, CastDataContainer
 from .BlockDataContainer import BlockDataContainer
 from .BlockGeometry import BlockGeometry
 from .framework import DataOrder
-from .framework import Partitioner
+from .Partitioner import Partitioner
